+++ tests/ranking.rs
@@ -486,6 +486,30 @@ mod ram_tests {
     }
 }
 
+mod spotis_tests {
+    use super::*;
+
+    #[test]
+    fn test_rank() -> Result<(), McdmError> {
+        let matrix = dmatrix![
+            2.9, 2.31, 0.56, 1.89;
+            1.2, 1.34, 0.21, 2.48;
+            0.3, 2.48, 1.75, 1.69
+        ];
+        let weights = dvector![0.25, 0.25, 0.25, 0.25];
+        let bounds = dmatrix![0.1, 3.0; 1.0, 2.48; 0.0, 1.9; 0.69, 3.1];
+        let criteria_types = mcdm::CriteriaType::from(vec![-1, 1, 1, -1])?;
+        let ranking = matrix.rank_spotis(&criteria_types, &weights, &bounds)?;
+        assert_relative_eq!(
+            ranking,
+            dvector![0.57089264, 0.69544822, 0.14071266],
+            epsilon = 1e-5
+        );
+
+        Ok(())
+    }
+}
+
 mod topsis_tests {
     use super::*;